+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
@@ -7,7 +7,9 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.mvc.handleradapter.HandlerAdapter;
+import webmvc.org.springframework.web.servlet.mvc.handleradapter.HandlerAdapters;
 import webmvc.org.springframework.web.servlet.mvc.handlermapping.HandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.handlermapping.HandlerMappings;
 import webmvc.org.springframework.web.servlet.view.JspView;
 
 public class DispatcherServlet extends HttpServlet {
