+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handlermapping/HandlerMappings.java
@@ -1,9 +1,8 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc.handlermapping;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.mvc.handlermapping.HandlerMapping;
 
 import java.util.ArrayList;
 import java.util.List;
